Implement truly horrible gif creation with ffmpeg
Now all the images are written to disk and then ffmpeg is called to
encode the images into a video and then a GIF. These files are then
removed.
There is an issue with the compression part of doing it all in memory
so this is a workaround until that is fixed.